[containers] ItemIndexFactor::hitCount should be 64 Bit

+++ include/sserialize/containers/ItemIndexFactory.h
@@ -40,7 +40,7 @@ private:
 	DataOffsetContainer m_dataOffsets;
 	IdToOffsetsType m_idToOffsets;
 	ItemIndexSizesContainer m_idxSizes;
-	std::atomic<uint32_t> m_hitCount;
+	std::atomic<uint64_t> m_hitCount;
 	bool m_checkIndex;
 	bool m_useDeduplication;
 	ItemIndex::Types m_type;
